+++ src/lib.rs
@@ -11,7 +11,7 @@
 //!
 //!```rust
 //!
-//! use gazetteer_entity_parser::{Gazetteer, ParserBuilder, EntityValue, ParsedValue, ResolvedValue};
+//! use gazetteer_entity_parser::{gazetteer, Gazetteer, ParserBuilder, EntityValue, ParsedValue, ResolvedValue};
 //!
 //! let mut gazetteer = Gazetteer::default();
 //! // We fill the gazetteer with artists, sorted by popularity
@@ -36,6 +36,15 @@
 //!     raw_value: "daniel brel".to_string(),
 //! });
 //!
+//! // Alternatively the gazetteer can be created with the `gazetteer!()` macro
+//! gazetteer = gazetteer!(
+//!     ("the rolling stones", "The Rolling Stones"),
+//!     ("the strokes", "The Strokes"),
+//!     ("the hives", "The Hives"),
+//!     ("jacques brel", "Jacques Brel"),
+//!     ("daniel brel", "Daniel Brel"),
+//! );
+//!
 //! // The Parser is then instantiated using a builder pattern. The ParserBuilder is instantiated
 //! // from a gazetteer and a decoding threshold, i.e. the minimal fraction of matched tokens for
 //! // a parsing to be possible. Additional methods allow to set the stop words of the Parser,
@@ -46,8 +55,11 @@
 //!     .n_stop_words(1)
 //!     .additional_stop_words(vec!["a".to_string(), "for".to_string()]).build().unwrap();
 //!
+//! // Maximum number of alternative resolved values to return
+//! let max_alternatives = 5;
+//!
 //! // Parse a sentence
-//! let parsed_stones = parser.run("I want to listen to the stones").unwrap();
+//! let parsed_stones = parser.run("I want to listen to the stones", max_alternatives).unwrap();
 //! assert_eq!(
 //!     parsed_stones,
 //!     vec![ParsedValue {
@@ -62,7 +74,7 @@
 //! );
 //!
 //! // Example with an ambiguity, where the artist with smaller rank is preferred
-//! let parsed_brel = parser.run("I want to listen to brel").unwrap();
+//! let parsed_brel = parser.run("I want to listen to brel", max_alternatives).unwrap();
 //! assert_eq!(
 //!     parsed_brel,
 //!     vec![ParsedValue {