+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 ### Changed
+- Add `max_alternatives` parameter to the `Parser::run` API [#39](https://github.com/snipsco/gazetteer-entity-parser/pull/39)
 - Add `alternatives` attribute in `ParsedValue` [#39](https://github.com/snipsco/gazetteer-entity-parser/pull/39)
 - Switch `matched_value` and `raw_value` [#39](https://github.com/snipsco/gazetteer-entity-parser/pull/39)
 - Group `resolved_value` and `matched_value` in a dedicated `ResolvedValue` object [#39](https://github.com/snipsco/gazetteer-entity-parser/pull/39)